백엔드 파일 크기 오버플로우 해결

+++ server/src/main/java/com/domi/ggmassetbackend/services/AssetService.java
@@ -325,7 +325,7 @@ public class AssetService {
             } catch (DomiException ignored) {}
         }
 
-        Integer fileSize = form.getFileSize();
+        Long fileSize = form.getFileSize();
         if (fileSize != null) {
             asset.setFileSize(fileSize > 0 ? fileSize : null);
         }
